\name{CNLTtsa-package}
\alias{CNLTtsa-package}
\alias{CNLTtsa}
\docType{package}
\title{
\packageTitle{CNLTtsa}
}
\description{
\packageDescription{CNLTtsa}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CNLTtsa}
\packageIndices{CNLTtsa}

The main routines of the package are \code{\link{cnlt.univ}} and \code{\link{cnlt.biv}} which perform the nondecimated complex-valued lifting transform
for univariate and bivariate time series, respectively; \code{\link{cnlt.spec}} computes spectral quantities of interest, for example the coherence and phase between two components of a bivariate series.

}
\author{
\packageAuthor{CNLTtsa}

Maintainer: \packageMaintainer{CNLTtsa}
}
\references{
Hamilton, J., Nunes, M. A, Knight, M. I. and Fryzlewicz, P. (2018) Complex-valued wavelet lifting and applications.
\emph{Technometrics}, \bold{60} (1), 48-60, DOI 10.1080/00401706.2017.1281846.\cr\cr
For related literature on the lifting methodology adopted in the technique, see\cr\cr
Knight, M. I. and Nason, G. P. (2009) A 'nondecimated' wavelet transform. \emph{Stat. Comput.}
\bold{19} (1), 1--16.\cr\cr
Knight, M. I., Nunes, M. A. and Nason, G. P. (2012) Spectral Estimation for Locally Stationary Time Series with Missing Observations. \emph{Stat. Comput.}
\bold{22} (4), 877--895.
}

\keyword{ package }
\seealso{
	\code{\link[CNLTreg]{cnlt.reg}}}
